+++ vector/src/main/java/im/vector/app/core/di/ScreenComponent.kt
@@ -69,6 +69,7 @@ import im.vector.app.features.roommemberprofile.devices.DeviceListBottomSheet
 import im.vector.app.features.roomprofile.RoomProfileActivity
 import im.vector.app.features.roomprofile.alias.detail.RoomAliasBottomSheet
 import im.vector.app.features.roomprofile.settings.historyvisibility.RoomHistoryVisibilityBottomSheet
+import im.vector.app.features.roomprofile.settings.joinrule.RoomJoinRuleBottomSheet
 import im.vector.app.features.settings.VectorSettingsActivity
 import im.vector.app.features.settings.devices.DeviceVerificationInfoBottomSheet
 import im.vector.app.features.share.IncomingShareActivity
@@ -157,6 +158,7 @@ interface ScreenComponent {
     fun inject(bottomSheet: RoomListQuickActionsBottomSheet)
     fun inject(bottomSheet: RoomAliasBottomSheet)
     fun inject(bottomSheet: RoomHistoryVisibilityBottomSheet)
+    fun inject(bottomSheet: RoomJoinRuleBottomSheet)
     fun inject(bottomSheet: VerificationBottomSheet)
     fun inject(bottomSheet: DeviceVerificationInfoBottomSheet)
     fun inject(bottomSheet: DeviceListBottomSheet)